requirements.txt
scipy==1.7.3
numpy==1.21.6
matplotlib==3.1.2
opencv_python==4.1.2.30
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0
onnx==1.10.1
onnxruntime-gpu==1.9.0
onnxsim==0.4.10